introduce Position type
The enum package was renamed to types.
The Move struct declaration was moved to types package.
A Position type was introduced to simplify data sharing between packages.
The code is still messy; the main area that needs rework is the tests.
I've realized that writing long test functions can screw refactoring up.
Additionally, the threefold repetition rule implementation must the addressed as soon as possible.