speed up move generation
Move generation speed was increased by ~4x using a legal move
generator. The issue with pseudo-legal move generation was due
to Go not supporting 'memcpy'-like functionality to directly copy
chunks of RAM. As a result, checking whether a move was legal and
then restoring the board was significantly slower than generating
only legal moves in the first place.